+ Add a dhtclient tool
The dhtclient tool is a simple command line client that can interact
with DHT nodes. All the main DhtClient requests are supported as command
line "commands".

The basic usage is: dhtclient -S nodes.xml <command> [arguments]

The commands are named after the DhtClient requests, but all in
lowercase and supporting shorter aliases (for example get -> g).

The help is a little weak for now, but fairly complete. If you type an
invalid command you will get a list of available commands and a brief
description of each. If you pass the wrong number of arguments to
a command, you will get a help message for that particular command, with
all the arguments it takes and their description. Unfortunately the
--help flag doesn't show any commands help (because the DhtTool class
need some work to enable that).