* dhtclient: Split reusable Command code to a different module
Make a better separation between the reusable code of the Command framework
and put it in its own Command module (which could be moved to ocean, or
maybe tango, in the future). The specific commands of the DHT client stays
in the Commands module.

Also a help command is added to get a list of available commands or
detailed help about any particular command.

Coding style is also adjusted to be more in line with the current code.

Note that this commit depends on tango.core.Array.map() function, which is
not present in upstream Tango yet.