* Rename dhtclient/DhtClientTool to dhtcli/DhtCli
The original name was a little ambiguous as there are plenty of DHT clients
and there is also another widespread library object called DhtClient (which
was alread cause some name clashes in the code).

dhtcli stands for DHT command line interface, but also can be interpreted
as a shortcut for the old name.