* Allow specifying the GC to use in the Makefile
Now it is possible to select which GC to use when compiling. This needs
some cooperation from the environment of the user to work. First you
need to have both versions of the Tango runtime compiled and using the
standard names libtango-dmd-<GC>.a and libtango-dmd-<GC>-dbg.a.

Then you have to modify your dmd.conf file and replace the -defaultlib
and -debuglib with something like this:

-defaultlib=tango-dmd-%D_GC% -debuglib=tango-dmd-%D_GC%-dbg

This will take replace %D_GC% with the contents of the environment
variable D_GC. Finally you have to add a nice default in your ~/.profile
if you want to be able to keep compiling random D files without the help
of a Makefile, something like this will do:

export D_GC=cdgc

If you want to use cdgc by default. Remember that for this change to
acknowledged by the shell, you have to re-login (or do the export in
each shell you're using).