* Don't prepend executable path to log files
When reading the location for log files from the config file, don't
prepend the executable path to them files.

Bare in mind that if you're not running the DHT node in the current
directory (which should be very rare), you need to update your config
file to include the complete path to the log files.