* cli: Move commands callbacks to methods
Using delegate literals was messing with the stack, because, even when
the "this" looked unused at first sight, it was used to get the vtable
as all methods in D (not marked as final) are virtual.

Other fixes could be applied (like making the methods final, or using
functions instead of methods) but this solution was chosen because there
already were some other callbacks as command methods. Also using
delegate literals with asynchronism is too fragile, and other problems
will arise sooner or later.