+ cli: Add support for "key formatters"
Now when specifying a key, besides the default behaviour of interpreting
them as an integer, it's possible to use a format specifier. Right now
there are 2 implemented:

i:key   Key is interpreted as an integer, just like the default (and
        previous) behaviour
h:key   Key is hashed via fnv1 algorithm

New format specifiers can be added in the future as needed.