* cli: Add format specifiers for --filter too
Make the option --filter accept format specifiers. s:filter is the same
as the default, it just use filter literally. i:filter and h:filter is
the same as for keys, the first is interpreted as an integer and the
last is first hashed via fvn1 and interpreted as an integer too (is
a "binary" filter, useful for looking for a particular ID inside
a value).