Allow stats log to be configured
As reported in #22, the stats log could not previously be configured (from
the application's config.ini). This can lead to problems where the default
values (which are defined in ocean.util.log.Stats) are unsuitable. For
example, in the logfiles node, where there are hundreds of channels, the
size of the lines being written to the stats log is quite large, and a
maximum file size of 10Mb (the default value) is not sufficient.