Fix bug with searching ranges within a single slot dir
The code previously assumed that *either* the current slot being scanned
is at the start of the range being searched or at the end of it, not both.
It is obviously possible for it to be both -- consider the case where the
search range is within a single slot, thus the slot being scanned is both
the first and the last slot in the search range.

Fixes #54.