Add unittest for LogRecord.getFirstBucket_()
The logic in this function is relatively complicated, and has been the
location of several bugs discovered over the last 6 months (including one
more while writing this unittest!), so it was regarded as a very worthwhile
candidate for unittesting.

This unittest requires the implementation of an "infrastructure" class --
DummyFileSystem. Once we settle on a better package layout for unittests,
this test could be moved out of this module.

Fixes #52.