Adapt periodics to reregister with epoll upon error
Previously, if an error occurred while handling a periodic, it would simply
be unregistered from epoll. This commit makes sure that periodics will
restart on error. An epoll instance is required for this, which is now
passed to the periodics' ctors.

The commit also updates some previously missing documentation of periodics'
ctors.