Combine periodic stats code for queue and dht
The stats writing code for the queue and dht was originally two completely
separate (although co-inspired) implementations. When the dht and queue
node projects were combined into swarmnodes, the stats code was, it seems,
only partially combined.

This commit unifies bits and pieces of code into the PeriodicStats base
class which were previously partially duplicated in the derived classes for
the dht and queue nodes.

Additionally, during the refactoring:
* PeriodicDhtStats.consoleOutput_() renamed extraConsoleInfo(), for clarity.
* Some members of PeriodicStats made private, as they are no longer needed
  by the sub-classes.