Add detection of out-of-range records while loading a .tcm file
Previously, out-of-range records (i.e. those not supported by the dhtnode's
specified hash range) were blindly loaded without any validation. This
commit adds this validation and two modes of behaviour upon finding an
invalid record key:
1. abort the program by throwing an exception
2. load the record but log a warning

The second behaviour is the default, and is compatible with the node's old
behaviour (plus the additional logging).

Another commit will add the ability to configure the behaviour of the node.

Part of #56