Add size of key/value length specifier to bytes written count
When a record is written to the dump file, in addition to the actual data
of the key and value, an 8-byte length specifier is written for each,
denoting how many bytes the following field (the key or value) contains.
Previously these length specifiers were not being counted in the stats
which track how much data is written to disk.