Fix loading of channel dumps without extra bytes at end
See #107.

The change in the dump file format introduced in sociomantic/swarm@2045e1f
inadvertently modified the loading of files in the old format. The loading
logic attempted to read one additional record from the file after the
actual number of records had been read, causing an IO error (and the node
to exit). Many dump files did not suffer from this problem as, due to
padding, sufficient extra bytes *did* exist at the end of the file.

This commit has three parts:
1. The file loading logic is fixed so that, for files in the old format,
only the exact number of records specified will be read from the file.
2. A unittest is added, to test the behaviour of the loadChannel() method
for files both with and without extra bytes at the end and in the old
versionless format and the new version 0 format.
3. For the purposes of unittesting, the ChannelLoader class had to be
adapted to allow a non-disk-based version to be created. Likewise, the
loadChannel() method of DumpManager is now static.