Add new NodeHashRange class for centralised hash range getting and setting
Previously the node's hash range was passed through from the top-level app
class down to the storage engines as a min & max hash. These values were
copied at each level where they're needed.

As part of #121, we need to be able to dynamically change the hash range of
a node, which means that the min & max hash values need to be adjusted at
all levels of the program. The hash range has thus been encapsulated in a class, a single instance of
which can be shared at all levels.

The new class also includes facilities for setting and clearing the hash
range, including adjusting the config file to save the new hash range.