Add responsibleForKey() method to DhtStorageChannels
The (2-line) function is copied from DhtStorageEngine. (I originally
thought of *moving* the method from DhtStorageEngine into
DhtStorageChannels, but DhtStorageEngine doesn't have a reference to the
DhtStorageChannels instance which owns it, so this would have been a more
complicated change, probably at the level of swarm. It doesn't seem worth
it for this minor change.)