Also dump empty channels
The previous behaviour (of both the node's dump thread, its shutdown
behaviour and the dump process) was to not dump empty channels (or to
delete them after dumping). In the case where a channel, at one point,
contained data (1 or more records) but no longer does, this causes problems
with alerts which are triggered when a dump file exceeds a certain age. We
decided that the simplest solution was to treat empty channels in the same
way as all others.

Fixes #185