Fix detection of in-progress dump files (for log message)
The logic previously checked for a files ending with ".dumping". However,
due to the use of mkostemp() (in swarmnodes.dht.memory.storage.DirectIO),
the file suffix of an in-progress dump file has an additional six
characters on the end. Simply checking that the file suffix *starts with*
".dumping" is enough to detect these files.