Do not write .backup file when dumping channels
Now that we are (since quite some time) dumping channel data into a
temporary file which is swapped with the actual dump file when complete,
there's no longer any need to also keep a backup file. (This essentially
served the same purpose as the temporary file we're using now.)

This also simplifies the creation of nagios alert scripts to check the
integrity of the dhtnodes' data directories.

The dumped channel data is being backed up anyway, by logrotate.

Fixes #189.