Remove all uses of deprecated Ocean's Trace
Remove all the uses of `ocean.util.log.Trace`, as it has been deprecated.
Uses of `Trace` have been replaced with `Stderr`, except for the start-up
message that is now printed to `Stdout`. In the future the uses of
`Stderr` should probably be switched to use the logging system instead.