mem: Add option to use regular I/O for dump/load
Using direct I/O is great for production systems, but it imposes a hard
restriction on the kind of filesystem where the node can read/write to,
making testing very complicated.

This commit adds an option to the dhtnode and dhtdump programs,
disable_direct_io, which does just that. When it's true, it just doesn't
use the O_DIRECT flag when opening files, using regular, page-cached, I/O.
This change is made deep down the code, so even when using regular I/O,
most of the code path is tested anyway.

The code in the dhtdump has to be rearranged a bit to be able to pass the
new config while creating the ChannelDumper (the creation was moved in the
DumpCycle from the constructor to the start() method).

To implement this we need a very recent feature in ocean, so the
submodule is advanced.

* submodules/ocean v1.17.0(64b78db)...v1.17.0+3(910e47a) (3 commits)
  > DirectIO: Add newFile() protected method to BufferedDirectReadFile
  > README: Adopt new release notes file
  > Fix cut & paste error in ClassFiller's LimitInit template

Fixes #210.