Remove temporary file if mkostemp(3) fails
Apparently mkostem(3) first opens the file and then tries to set it to
use O_DIRECT, so even when mkostemp(3) fails, it can leave a temporary
file behind. This is not good, as it pollutes the filesystem
unnecessarily. Since mkostemp(3) is indirectly called by the open() call
for a file, it is completely unexpected to the caller that after open()
failed, a file could be left behind, so we handle the removal of the
file locally where mkostemp(3) is used.

Fixes #215, improves #216.