Silently ignore the deleted directory
While scanning for dump files, a warning was printed about the "deleted"
directory, which is a completely normal situation. The check for
subdirectories in the data directory no longer warns about the "deleted"
directory. To make sure the deleted folder is the same, we use the
absolute paths to compare now.