Use a better suffix for files being dump
The suffix appended by the temporary file dump class was not in the
control of the class needing to actually rotate or scan the files, the
NewFileSuffix was somehow useless, as another suffix was appended to it.

This commit replaces the use of mkostemp(3) with mkostemps(3) to be able
to control the suffix. Now the dump files really have a suffix as specified
by NewFileSuffix, apart from the variable part to ensure the file's
uniqueness. This effectively changes the dump file pattern from
CHAN.tcm.dumpingXXXXXX to CHAN.tcm.XXXXXX.dumping, which also makes it more
readable.