Create logfiles/dht-specific connection handlers
This is a first step towards being able to remove the supported commands
infrastructure, which is no longer needed. This commit changes the
behaviour of the nodes when receiving a valid but not supported command.
Traditionally this would result in a NotSupported error but now the node
will simply break the connection with the client, in exactly the same way
as if a completely unknown command code were sent.

Part of #224.