Make shared resources mixin template for connection handlers
Currently the shared resources for the dhtnode and the logfilesnode are
identical. They may diverge in the future (even in the current set of
shared resources, not all resource types are needed by both types of node)
but while this is the case, it's better if the code is shared.