dhtnode: when loading from disk, check validity of keys, ignore invalid
As the dht record keys are transmitted and stored in the form of char[]s,
there is the possibility of invalid keys -- either strings containing non-
hexadecimal characters or strings of length != 16.

This commit only partially addresses this issue, as it exists potentially
throughout the whole system, anywhere a key is read from an external
source. This commit fixes the problem at the point where keys are read
tcm files while the dhtnode starts up.

Fixes #239.