Make dhtdump perform node handshake at startup
Previously dhtdump did not perform a node handshake, as it would fail
because it's only contacting a single node and would thus not have a
complete hash range. The behaviour of the handshake in swarm has now
changed, such that connecting to a single node is now possible (requests
may still be sent, even if the complete hash range is not covered), but
the handshake is required by all apps as the client will refuse to send
requests to any node whose API version is unknown or mismatched.

Fixes #254.