Remove invalid command handling code
The swarm nodes previously would write an InvalidRequest status code to the
client when an unknown request code is received. However, immediately
afterwards, the base connection handler in swarm throws and exception,
causing the socket to be closed. The error code is thus never sent and has
no effect.

See swarmnodes#166.
See swarmnodes#283.