Move records handled counting into storage engine
This simplifies the request handling classes and moves the stats counting
into a single, central location. Due to the fact that recordsHandled() is
a method of the top-level INodeInfo interface, which is implemented by
Node in swarm, a slight work-around is required: passing a delegate to the
storage channels.

Note that now records pushed to multiple channels via PushMulti will be
counted once per channel. Previously these were only counted once.

Fixes #57.