Delete channel dump files after all channels were successfully loaded
This avoids sending duplicate records to the queue consumers when
restarting the queue node after it was killed.
Previously after loading a channel dump file it was kept and deleted on a
graceful shutdown on SIGINT or SIGTERM. If the queue node was killed for
example by the OOM killer, the dump files remained and were loaded again on
the next start.