Fix not resetting overflow data file when removing the last channel
When removing a channel, the channel metadata, containing the number of
records and bytes in the channel, were reset to 0 before subtracting them
from the total amounts so they were actually not subtracted. As a
consequence the overflow data file was not deleted after removing the last
channel, even thought the disk overflow was empty. One way how this
inconsistency manifests is that, if the node is restarted in this
condition and a record is then pushed to the disk overflow, an enforcement
fails that for an empty disk overflow the data file should be empty.

The order of resetting and subtracting is wrong, reset the channel metadata
after its amounts have been subtracted from the total.