Replace PosixFile.unlink() with remove() to really delete the file
unlink() only removes the entry in the file system but does not deallocate
the storage space used by the file until the file is closed. However, the
PosixFile invariant prevents PosixFile.close() from being called after
PosixFile.unlink(), so the file keeps being allocated until it is closed
automatically when the program exits.
Fix this by replacing PosixFile.unlink() with remove(), which also closes
the file.