Fix overflow-only channels not created on startup
Previously only channels with a memory dump file were created on startup.
It is possible that a channel has an empty memory queue but disk overflow
data, such a channel will have no memory dump file so it won't be created
on startup.
Fix this error by doing a `getCreate()` for all channels in the disk
overflow.