Minimize the disk overflow data file size
The data file starts with an 8-byte identifier string, and the records
follow that string. After records have been popped, there may be a
contiguous range of already popped records between the identifier string
and the first record in the data file that hasn't been popped yet. If such
a range exists then DiskOverflow.flush() will minimize the size of the data
file by truncating it from the beginning.





# Conflicts:
#	src/dmqnode/storage/engine/DiskOverflow.d

# Conflicts:
#	src/dmqnode/storage/engine/DiskOverflow.d